fix duty cycle datatype
Should be float, as it is a normalised value and not the actual PWM value.

+++ rpi/src/System.cpp
@@ -161,7 +161,7 @@ float System::getLux(size_t id)
     }
 }
 
-int System::getDutyCycle(size_t id)
+float System::getDutyCycle(size_t id)
 {
     boost::shared_lock<boost::shared_mutex> lock(mutex_);
     try
